// src/type.ts
export interface IElectiveBatch {
  bybz: string; // '0';
  beginTime2: string; // '2024-10-18 20:30:00';
  endTime2: string; // '2024-10-23 08:00:00';
  beginTime3: string; // '2024-11-01 20:30:00';
  endTime3: string; // '2024-11-06 08:00:00';
  code: string; // '23d4bd1b8188422bbbba6606989b20b0';
  name: string; // '2024-2025学年冬季学期';
  canSelect: string; // '1';
  schoolTerm: string; // '2024-2025-2';
  beginTime: string; // '2024-10-11 20:30:00';
  endTime: string; // '2024-10-16 08:00:00';
  tacticCode: string; // '01';
  typeCode: string; // '02';
  typeName: string; // '正选';
  isConfirmed: string; // '0';
  schoolTermName: string; // '2024-2025学年冬季学期';
  weekRange: string; // '1-10周';
  canSelectBook: string; // '0';
  canDeleteBook: string; // '0';
  multiCampus: string; // '1';
  multiTeachCampus: string; // '1';
  noCheckTimeConflict: string; // '0';
}